backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "livekit-agents[cartesia,deepgram,google,silero,turn-detector]~=1.0",
    "livekit-api>=1.0.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
